@@ -1,5 +1,4 @@
 ﻿using System;
-using System.Collections;
 using System.Collections.Generic;
 using System.IO;
 
@@ -18,6 +17,7 @@ namespace Pester
                 FailedCount = block.FailedCount,
                 PassedCount = block.PassedCount,
                 SkippedCount = block.SkippedCount,
+                InconclusiveCount = block.InconclusiveCount,
                 NotRunCount = block.NotRunCount,
                 TotalCount = block.TotalCount,
                 ErrorRecord = block.ErrorRecord ?? new List<object>(),
@@ -47,7 +47,13 @@ namespace Pester
             };
         }
 
-        public string Type { get; set; }
+        public string Name { get => ToStringConverter.ContainerItemToString(Type, Item); }
+        private string _type = Constants.File;
+        public string Type
+        {
+            get => _type;
+            set => SetContainerType(ref _type, value);
+        }
         public object Item { get; set; }
         public object Data { get; set; }
         public List<Block> Blocks { get; set; } = new List<Block>();
@@ -56,6 +62,7 @@ namespace Pester
         public int FailedCount { get; set; }
         public int PassedCount { get; set; }
         public int SkippedCount { get; set; }
+        public int InconclusiveCount { get; set; }
         public int NotRunCount { get; set; }
         public int TotalCount { get; set; }
         public List<object> ErrorRecord { get; set; } = new List<object>();
@@ -74,6 +81,16 @@ namespace Pester
         public override string ToString()
         {
             return ToStringConverter.ContainerToString(this);
+        }
+
+        internal static void SetContainerType(ref string property, string value)
+        {
+            property = value.ToLower() switch
+            {
+                "file" => Constants.File,
+                "scriptblock" => Constants.ScriptBlock,
+                _ => throw new ArgumentOutOfRangeException("value", $"Type must be '{Constants.File}' or '{Constants.ScriptBlock}'"),
+            };
         }
     }
 }
