@@ -46,7 +46,7 @@ namespace Pester
             }
         }
 
-        public OutputConfiguration() : base("Output configuration")
+        public OutputConfiguration() : base("Options to customize the console output generated by Pester.")
         {
             Verbosity = new StringOption("The verbosity of output, options are None, Normal, Detailed and Diagnostic.", "Normal");
             StackTraceVerbosity = new StringOption("The verbosity of stacktrace output, options are None, FirstLine, Filtered and Full.", "Filtered");
