@@ -48,7 +48,7 @@ namespace Pester.Tracing
         // keyed as path -> line:column -> CodeCoveragePoint
         public Dictionary<string, Dictionary<string, List<CodeCoveragePoint>>> Hits { get; } = new Dictionary<string, Dictionary<string, List<CodeCoveragePoint>>>(StringComparer.OrdinalIgnoreCase);
 
-        public void Trace(string message, IScriptExtent extent, ScriptBlock _, int __)
+        public void Trace(string message, IScriptExtent extent, ScriptBlock _, int __, string ___, string ____)
         {
             if (_debug && extent?.File != null && CultureInfo.InvariantCulture.CompareInfo.IndexOf(extent.File, _debugFile, CompareOptions.OrdinalIgnoreCase) >= 0)
             {
@@ -102,10 +102,10 @@ namespace Pester.Tracing
 
 #pragma warning disable IDE0060
         // Profiler v3.1 compatible overload
-        public void Trace(IScriptExtent extent, ScriptBlock _, int __) => Trace(null, extent, _, __);
+        public void Trace(IScriptExtent extent, ScriptBlock scriptBlock, int level) => Trace(null, extent, scriptBlock, level, null, null);
 
         // Profiler v4 compatible overload
-        public void Trace(IScriptExtent extent, ScriptBlock _, int __, string ___, string ____) => Trace(null, extent, _, __);
+        public void Trace(IScriptExtent extent, ScriptBlock scriptBlock, int level, string functionName, string moduleName) => Trace(null, extent, scriptBlock, level, functionName, moduleName);
 #pragma warning restore IDE0060
     }
 }
